fix(stella-cli): an invoked skill announces itself and records a use
Typing `/some-skill` inlined the skill's body into the prompt and told
nobody. `extensions::expand` returned prompt text; no event went on the
channel, so the transcript had no `✦ skill` row for it — and, worse, the
`skill_usage` table counted nothing, because that table is written from a
re-run of *auto* selection and selection never runs for a `/slug`. Skill
appraisal reads exactly that table before retiring a skill for non-use, so
a skill invoked daily looked unused. The same phantom-non-usage failure as
#1872, one channel over.

`AgentEvent::SkillInjected` gains a `SkillTrigger` — `Auto` or `Command`,
rendered as SPEC 6.3's `auto|/cmd` head. Both arms have a producer in this
PR (#4180's rule): auto-selection in `RecalledBlock::telemetry_events`, and
the invocation through `OpeningRecall::note_invoked_skill`, which the
interactive loop and the deck both call once recall exists to carry it. The
field is `#[serde(default)]` to `Auto`, which is what a journal written
before this actually meant, so an older stream still decodes.

`skill_usage` learns the second channel too. `stamp_and_record_skill_usage`
takes the invoked slugs, read back off the events rather than kept beside
them, so what the transcript shows as invoked and what the table counts as
invoked are the same list by construction. A skill both selected and
invoked records one row, under selection's reason.

Witnesses, at four levels:

- `an_invoked_skill_reaches_the_channel_as_a_command` — the event exists
  and carries `Command`. Checked the flip artisanally: setting the trigger
  to `Auto` fails it with `left: Auto, right: Command`.
- `a_skill_invoked_by_name_records_a_use_selection_would_have_missed` —
  the usage row lands on a prompt that selects nothing, asserted as a
  control first so the row cannot be selection's.
- `a_skill_both_selected_and_invoked_records_one_use` — no double count.
- `an_invoked_skill_renders_the_command_trigger` — the head reads `/cmd`.
  The renderer's trigger column was a literal `"auto"`.
- `an_invoked_skill_is_handed_back_to_be_recorded` — `expansion` reports
  the skill, and reports none for a command or an agent persona.

Two structural notes. `CustomExtensions::expand` is now `#[cfg(test)]`:
every live caller needs the skill as well as the prompt, so the lint was
right that nothing shipping calls it, and `#[cfg(test)]` says that rather
than `#[allow(dead_code)]` asserting the lint is wrong. And the change put
`extensions.rs` at 1549 lines, so its test module moved to
`extensions/tests.rs` — no baseline entry, per AGENTS.md.

No test is deleted: `extensions.rs`'s tests are moved into
`extensions/tests.rs` unchanged. `docs/wire/` is regenerated.

Closes #5232
