+++ crates/stella-protocol/src/event/kind.rs
@@ -1047,20 +1047,29 @@ pub enum AgentEvent {
         task_id: Option<crate::TaskId>,
     },
     /// A workspace skill entered this turn's context — one event per skill
-    /// the steering section actually carried.
+    /// that actually reached the prompt, by either channel.
     ///
-    /// Emitted after the section's own token budget has made its cut, so
-    /// `tokens` is what the prompt paid rather than what selection ranked:
-    /// selection routinely scores more skills than the section fits, and an
-    /// event per *candidate* would bill the turn for text it never sent.
+    /// For [`SkillTrigger::Auto`] it is emitted after the steering section's
+    /// own token budget has made its cut, so `tokens` is what the prompt paid
+    /// rather than what selection ranked: selection routinely scores more
+    /// skills than the section fits, and an event per *candidate* would bill
+    /// the turn for text it never sent. For [`SkillTrigger::Command`] there is
+    /// no cut to wait for — the expansion inlines the skill's whole body.
     SkillInjected {
         /// The skill's slug, as its `SKILL.md` frontmatter names it.
         name: String,
         /// The skill's own one-line description.
         summary: String,
         /// What this skill's rendered block cost, estimated over the exact
-        /// bytes the section carried.
+        /// bytes the prompt carried.
         tokens: u32,
+        /// Which channel put it there.
+        ///
+        /// Absent on a journal written before #5232, where the only producer
+        /// was auto-selection — so the default is the value those events
+        /// meant.
+        #[serde(default = "SkillTrigger::auto")]
+        trigger: SkillTrigger,
     },
     /// A context block first became eligible to enter the prompt (spec §4).
     /// The birth record that makes the per-step manifest an index over the fold.
