+++ crates/stella-protocol/tests/wire_contract.rs
@@ -481,6 +481,7 @@ fn every_nested_vocabulary_is_fully_sampled() {
         ("ServiceTier", all_service_tiers().len()),
         ("Verbosity", all_verbosities().len()),
         ("FileChangeKind", all_file_change_kinds().len()),
+        ("SkillTrigger", all_skill_triggers().len()),
         ("ProofTree", all_proof_trees().len()),
         ("LadderRung", all_ladder_rungs().len()),
         ("FlipOutcome", all_flip_outcomes().len()),
@@ -539,6 +540,7 @@ fn every_nested_vocabulary_is_fully_sampled() {
         "ServiceTier",
         "Verbosity",
         "FileChangeKind",
+        "SkillTrigger",
         "ProofTree",
         "LadderRung",
         "FlipOutcome",
